+++ src/constants.ts
@@ -3,3 +3,8 @@ export const $effect = Symbol.for('effect');
 export const $destroySignal = Symbol.for('destroySignal');
 export const $createEffect = Symbol.for('createEffect');
 export const $destroyEffect = Symbol.for('destroyEffect');
+
+export const Value = 'value';
+export const Mute = 'mute';
+export const Unmute = 'unmute';
+export const Destroy = 'destroy';