+++ src/createSignal.link.spec.ts
@@ -1,25 +1,23 @@
 import {on} from '@spearwolf/eventize';
-import {assertEffectsCount, assertSignalsCount} from './assert-helpers.js';
-import {signal} from './decorators.js';
 import {
-  Connection,
-  connect,
-  createSignal,
-  destroyObjectSignals,
-  destroySignal,
-  touch,
-  unconnect,
-} from './index.js';
+  assertEffectsCount,
+  assertLinksCount,
+  assertSignalsCount,
+} from './assert-helpers.js';
+import {Destroy, Mute, Unmute, Value} from './constants.js';
+import {createSignal, destroySignal, link, touch, unlink} from './index.js';
 
 describe('connect signals', () => {
   beforeEach(() => {
     assertEffectsCount(0, 'beforeEach');
     assertSignalsCount(0, 'beforeEach');
+    assertLinksCount(0, 'beforeEach');
   });
 
   afterEach(() => {
     assertEffectsCount(0, 'afterEach');
     assertSignalsCount(0, 'afterEach');
+    assertLinksCount(0, 'afterEach');
   });
 
   it('connect a signal with another signal (naive version)', () => {
@@ -53,7 +51,7 @@ describe('connect signals', () => {
     destroySignal(sigA, sigB, sigC);
   });
 
-  it('connect a signal with another signal (this time with the connect() api)', () => {
+  it('connect a signal with another signal (this time with the link() api)', () => {
     const {get: sigA, set: setA} = createSignal(1);
     const {get: sigB, set: setB} = createSignal(-1);
 
@@ -63,11 +61,11 @@ describe('connect signals', () => {
     expect(sigA()).toBe(1);
     expect(sigB()).toBe(-1);
 
-    const con = connect(sigA, sigB);
+    const con = link(sigA, sigB);
 
     on(con, {
-      [Connection.Value]: valueMock,
-      [Connection.Destroy]: destroyMock,
+      [Value]: valueMock,
+      [Destroy]: destroyMock,
     });
 
     // when we create a connection between two signals, the value of the source signal is immediately written to the target signal !!
@@ -75,9 +73,9 @@ describe('connect signals', () => {
     expect(sigA()).toBe(1);
     expect(sigB()).toBe(1);
 
-    expect(valueMock).toHaveBeenCalledTimes(1);
-    expect(valueMock).toHaveBeenCalledWith(1);
-    valueMock.mockClear();
+    expect(valueMock).toHaveBeenCalledTimes(0);
+    // expect(valueMock).toHaveBeenCalledWith(1);
+    // valueMock.mockClear();
 
     setB(100);
 
@@ -110,68 +108,6 @@ describe('connect signals', () => {
     destroySignal(sigA, sigB);
   });
 
-  it('connect a signal with another object signal', () => {
-    const {get: sigA, set: setA} = createSignal(1);
-
-    class Foo {
-      @signal() accessor b = -1;
-    }
-
-    const foo = new Foo();
-
-    expect(sigA()).toBe(1);
-    expect(foo.b).toBe(-1);
-
-    connect(sigA, [foo, 'b']);
-
-    expect(sigA()).toBe(1);
-    expect(foo.b).toBe(1);
-
-    foo.b = 100;
-
-    expect(foo.b).toBe(100);
-
-    setA(2);
-
-    expect(sigA()).toBe(2);
-    expect(foo.b).toBe(2);
-
-    destroySignal(sigA);
-    destroyObjectSignals(foo);
-  });
-
-  it('connect a signal with an object method', () => {
-    const {get: sigA, set: setA} = createSignal(1);
-
-    class Foo {
-      b: (val: number) => void = jest.fn();
-      c: (val: string) => void = jest.fn();
-    }
-
-    const foo = new Foo();
-
-    expect(sigA()).toBe(1);
-    expect(foo.b).toHaveBeenCalledTimes(0);
-
-    connect(sigA, [foo, 'b']);
-    // connect(sigA, [foo, 'c']);
-
-    expect(sigA()).toBe(1);
-    expect(foo.b).toHaveBeenCalledWith(1);
-    expect(foo.b).toHaveBeenCalledTimes(1);
-
-    expect(connect(sigA, [foo, 'b'])).toBe(connect(sigA, [foo, 'b']));
-    expect(foo.b).toHaveBeenCalledTimes(1);
-
-    setA(2);
-
-    expect(sigA()).toBe(2);
-    expect(foo.b).toHaveBeenCalledTimes(2);
-    expect(foo.b).toHaveBeenCalledWith(2);
-
-    destroySignal(sigA);
-  });
-
   it('connect a signal with a function', () => {
     const {get: sigA, set: setA} = createSignal(1);
 
@@ -180,10 +116,10 @@ describe('connect signals', () => {
     expect(sigA()).toBe(1);
     expect(bMock).toHaveBeenCalledTimes(0);
 
-    connect(sigA, bMock);
+    link(sigA, bMock);
     expect(bMock).toHaveBeenCalledWith(1);
 
-    expect(connect(sigA, bMock)).toBe(connect(sigA, bMock));
+    expect(link(sigA, bMock)).toBe(link(sigA, bMock));
     expect(bMock).toHaveBeenCalledTimes(1);
 
     setA(2);
@@ -194,122 +130,15 @@ describe('connect signals', () => {
     destroySignal(sigA);
   });
 
-  it('connect an object signal with another signal', () => {
-    class Foo {
-      @signal() accessor a = 1;
-    }
-
-    const foo = new Foo();
-
-    const {get: b} = createSignal(-1);
-
-    expect(foo.a).toBe(1);
-    expect(b()).toBe(-1);
-
-    connect([foo, 'a'], b);
-
-    expect(foo.a).toBe(1);
-    expect(b()).toBe(1);
-
-    foo.a = 2;
-
-    expect(foo.a).toBe(2);
-    expect(b()).toBe(2);
-
-    destroySignal(b);
-    destroyObjectSignals(foo);
-  });
-
-  it('connect an object signal with another object signal', () => {
-    class Foo {
-      @signal() accessor a = 1;
-      @signal() accessor b = -1;
-    }
-
-    const foo = new Foo();
-
-    expect(foo.a).toBe(1);
-    expect(foo.b).toBe(-1);
-
-    connect([foo, 'a'], [foo, 'b']);
-
-    expect(foo.a).toBe(1);
-    expect(foo.b).toBe(1);
-
-    foo.a = 2;
-
-    expect(foo.a).toBe(2);
-    expect(foo.b).toBe(2);
-
-    destroyObjectSignals(foo);
-  });
-
-  it('connect an object signal with an object method', () => {
-    class Foo {
-      @signal() accessor a = 1;
-
-      b: (val: number) => void = jest.fn();
-    }
-
-    const foo = new Foo();
-
-    expect(foo.a).toBe(1);
-    expect(foo.b).toHaveBeenCalledTimes(0);
-
-    connect([foo, 'a'], [foo, 'b']);
-
-    expect(foo.b).toHaveBeenCalledWith(1);
-    expect(foo.b).toHaveBeenCalledTimes(1);
-
-    expect(connect([foo, 'a'], [foo, 'b'])).toBe(
-      connect([foo, 'a'], [foo, 'b']),
-    );
-    expect(foo.b).toHaveBeenCalledTimes(1);
-
-    foo.a = 2;
-
-    expect(foo.a).toBe(2);
-    expect(foo.b).toHaveBeenCalledTimes(2);
-    expect(foo.b).toHaveBeenCalledWith(2);
-
-    destroyObjectSignals(foo);
-  });
-
-  it('connect an object signal with a function', () => {
-    class Foo {
-      @signal() accessor a = 1;
-    }
-
-    const foo = new Foo();
-
-    const bMock = jest.fn();
-
-    expect(foo.a).toBe(1);
-    expect(bMock).toHaveBeenCalledTimes(0);
-
-    connect([foo, 'a'], bMock);
-    expect(bMock).toHaveBeenCalledWith(1);
-
-    expect(connect([foo, 'a'], bMock)).toBe(connect([foo, 'a'], bMock));
-    expect(bMock).toHaveBeenCalledTimes(1);
-
-    foo.a = 2;
-
-    expect(foo.a).toBe(2);
-    expect(bMock).toHaveBeenCalledWith(2);
-
-    destroyObjectSignals(foo);
-  });
-
-  it('a connection between two points is a singleton and cannot be created more than once', () => {
+  it('a link between two points is a singleton and cannot be created more than once', () => {
     const {get: sigA, set: setA} = createSignal(1);
     const {get: sigB} = createSignal(-1);
 
     expect(sigA()).toBe(1);
     expect(sigB()).toBe(-1);
 
-    const con0 = connect(sigA, sigB);
-    const con1 = connect(sigA, sigB);
+    const con0 = link(sigA, sigB);
+    const con1 = link(sigA, sigB);
 
     expect(con0).toBe(con1);
 
@@ -317,7 +146,7 @@ describe('connect signals', () => {
 
     expect(sigB()).toBe(666);
 
-    unconnect(sigA, sigB);
+    unlink(sigA, sigB);
 
     setA(42);
 
@@ -327,7 +156,7 @@ describe('connect signals', () => {
     destroySignal(sigA, sigB);
   });
 
-  it('a connection should be pauseable (mute/unmute)', () => {
+  it('a link should be pauseable (mute/unmute)', () => {
     const {get: sigA, set: setA} = createSignal(1);
     const {get: sigB} = createSignal(-1);
 
@@ -337,11 +166,11 @@ describe('connect signals', () => {
     expect(sigA()).toBe(1);
     expect(sigB()).toBe(-1);
 
-    const con = connect(sigA, sigB);
+    const con = link(sigA, sigB);
 
     on(con, {
-      [Connection.Mute]: muteMock,
-      [Connection.Unmute]: unmuteMock,
+      [Mute]: muteMock,
+      [Unmute]: unmuteMock,
     });
 
     expect(con.isMuted).toBe(false);
@@ -380,7 +209,6 @@ describe('connect signals', () => {
     expect(muteMock).toHaveBeenCalledTimes(2);
     expect(unmuteMock).toHaveBeenCalledTimes(1);
 
-    unconnect(sigA, sigB);
     destroySignal(sigA, sigB);
   });
 
@@ -389,13 +217,13 @@ describe('connect signals', () => {
     const {get: sigB} = createSignal(-1);
     const {get: sigC} = createSignal(-1);
 
-    const con0 = connect(sigA, sigB);
-    const con1 = connect(sigA, sigC);
+    const con0 = link(sigA, sigB);
+    const con1 = link(sigA, sigC);
 
     const destroyMock = jest.fn();
 
-    on(con0, Connection.Destroy, destroyMock);
-    on(con1, Connection.Destroy, destroyMock);
+    on(con0, Destroy, destroyMock);
+    on(con1, Destroy, destroyMock);
 
     setA(666);
 
@@ -433,34 +261,34 @@ describe('connect signals', () => {
     setB(-1);
 
     expect(callingB).toHaveBeenCalledTimes(1);
-    expect(callingB).toBeCalledWith(-1);
+    expect(callingB).toHaveBeenCalledWith(-1);
 
-    const con = connect(sigA, sigB);
+    const con = link(sigA, sigB);
 
-    on(con, Connection.Value, valueMock);
+    on(con, Value, valueMock);
 
     expect(callingB).toHaveBeenCalledTimes(2);
-    expect(valueMock).toHaveBeenCalledTimes(1);
+    expect(valueMock).toHaveBeenCalledTimes(0);
 
     setA(666);
 
     expect(sigA()).toBe(666);
     expect(sigB()).toBe(666);
     expect(callingB).toHaveBeenCalledTimes(3);
-    expect(callingB).toBeCalledWith(666);
-    expect(valueMock).toHaveBeenCalledTimes(2);
-    expect(valueMock).toBeCalledWith(666);
+    expect(callingB).toHaveBeenCalledWith(666);
+    expect(valueMock).toHaveBeenCalledTimes(1);
+    expect(valueMock).toHaveBeenCalledWith(666);
 
     setA(666);
 
     expect(callingB).toHaveBeenCalledTimes(3);
-    expect(valueMock).toHaveBeenCalledTimes(2);
+    expect(valueMock).toHaveBeenCalledTimes(1);
 
     con.touch();
 
     expect(sigB()).toBe(666);
     expect(callingB).toHaveBeenCalledTimes(4);
-    expect(valueMock).toHaveBeenCalledTimes(3);
+    expect(valueMock).toHaveBeenCalledTimes(2);
 
     destroySignal(sigA, sigB);
   });
@@ -479,8 +307,8 @@ describe('connect signals', () => {
     expect(callingB).toHaveBeenCalledTimes(0);
     expect(callingC).toHaveBeenCalledTimes(0);
 
-    connect(sigA, sigB);
-    connect(sigA, sigC);
+    link(sigA, sigB);
+    link(sigA, sigC);
 
     expect(callingB).toHaveBeenCalledTimes(1);
     expect(callingC).toHaveBeenCalledTimes(1);
@@ -492,10 +320,10 @@ describe('connect signals', () => {
     expect(sigC()).toBe(666);
 
     expect(callingB).toHaveBeenCalledTimes(2);
-    expect(callingB).toBeCalledWith(666);
+    expect(callingB).toHaveBeenCalledWith(666);
 
     expect(callingC).toHaveBeenCalledTimes(2);
-    expect(callingC).toBeCalledWith(666);
+    expect(callingC).toHaveBeenCalledWith(666);
 
     setA(666);